// frontend/src/app/model/objectives/objective.ts
export interface Objective {
    populationGroup: string;
    chronicDisease: string;
    groupOfChronicDiseases: string;
    disease: string;
    objective: string;
    range: string;
    testOrQuestionnaire: string;
    specific: string;
    measurable: string;
    achievable: string;
    relevant: string;
    temporal: string;
    smartObjective: string;
}